Move SyntheticSourceRollingUpgradeIT to the new test framework
